app/requirements.txt
flask
flask-cors
requests
openai
Pillow
langchain-community
langchain
pymongo
bs4
tiktoken
gradio
lxml
argparse
unstructured
